requirements.txt
python-dotenv
pyTelegramBotAPI
groq
pypdf
fpdf
markdown
xhtml2pdf